requirements.txt
PyGithub
python-dotenv
tenacity